backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi-mcp>=0.4.0",
    "alembic>=1.15",
    "pgvector>=0.4",
    "bcrypt>=4.0",
    "pyjwt>=2.0",
    "openai>=1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aiosqlite>=0.21",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that call external services (OpenAI, etc.)",
]

[dependency-groups]
dev = [
    "ty>=0.0.26",
]
